refactor: Rename experiments folder to modelSessions
