+++ DashAI/front/src/components/models/AddModelDialog.jsx
@@ -18,8 +18,8 @@ import { Close as CloseIcon } from "@mui/icons-material";
 import { useSnackbar } from "notistack";
 import FormSchemaWithSelectedModel from "../shared/FormSchemaWithSelectedModel";
 import FormSchemaContainer from "../shared/FormSchemaContainer";
-import OptimizationTableSelectOptimizer from "../experiments/OptimizationTableSelectOptimizer";
-import ModelsTableSelectMetric from "../experiments/ModelsTableSelectMetric";
+import OptimizationTableSelectOptimizer from "./modelSession/OptimizationTableSelectOptimizer";
+import ModelsTableSelectMetric from "./modelSession/ModelsTableSelectMetric";
 import useSchema from "../../hooks/useSchema";
 import { generateSequentialName } from "../../utils/nameGenerator";
 import { createRun } from "../../api/run";
