+++ DashAI/front/src/components/models/RunCard.jsx
@@ -37,8 +37,8 @@ import { getRunStatus } from "../../utils/runStatus";
 import RunResults from "./RunResults";
 import FormSchemaWithSelectedModel from "../shared/FormSchemaWithSelectedModel";
 import FormSchemaContainer from "../shared/FormSchemaContainer";
-import OptimizationTableSelectOptimizer from "../experiments/OptimizationTableSelectOptimizer";
-import ModelsTableSelectMetric from "../experiments/ModelsTableSelectMetric";
+import OptimizationTableSelectOptimizer from "./modelSession/OptimizationTableSelectOptimizer";
+import ModelsTableSelectMetric from "./modelSession/ModelsTableSelectMetric";
 import useSchema from "../../hooks/useSchema";
 import { updateRunParameters, getRunOperationsCount } from "../../api/run";
 import RetrainConfirmDialog from "./RetrainConfirmDialog";
