+++ DashAI/front/src/components/models/modelSession/metrics/MetricsSelector.jsx
