+++ DashAI/front/src/components/models/modelSession/metrics/SplitColumn.jsx
